The tracker also has a key pair now to verify the user
The tracker on every startup now generates a key pair to verify the user.
The tracker sends its public key which is then used by the client to generate
a common signing key. The client signs its public key with the shared key,
and the tracker verifies the signature. If it's not valid the connection
is dropped.